Added new component of modal
